:sparkles::truck: Link other components into main and organize files a bit
