+++ Cargo.toml
@@ -14,7 +14,11 @@ members = [
 ]
 
 [dependencies]
-minifb = "0.28"
+bus = { version = "0.1.0", path = "./bus"}
+cpu = { version = "0.1.0", path = "./cpu"}
+ppu = { version = "0.1.0", path = "./ppu"}
+apu = { version = "0.1.0", path = "./apu"}
+
 eframe = "0.33.2"
 egui = "0.33.2"
 egui_extras = "0.33.2"
